feat: moved config folder to the backend, since its only related to him. fixed backed initialization ( pipe creation )
