+++ apps/backend/commands/commands.c
@@ -1,8 +1,52 @@
 #include "./commands.h"
 #include "../../../shared/helpers/helpers.h"
-#include "../../../shared/config/config.h"
+#include "../config/config.h"
 #include "../../../users_lib.h"
+/*
+ *
+// TODO: i think we should move the items list to a command !
+//  it was necessary for the first meta!
+void load_items_from_file(char *filename, Item *items)
+{
+    FILE *fp;
+    char *line = NULL;
+    size_t len = 0;
+    ssize_t read;
+    int i = 0;
 
+    fp = fopen(filename, "r");
+    if (fp == NULL)
+        exit(EXIT_FAILURE);
+
+    while ((read = getline(&line, &len, fp)) != -1)
+    {
+        Item it;
+        // the %20s is to avoid buffer overflows
+        sscanf(line, "%20s %20s %20s %d %d %d %20s %20s", it.identifier, it.name, it.category, &it.duration, &it.current_value, &it.buy_now_value, it.seller_name, it.bidder_name);
+        items[i++] = it;
+    }
+
+    fclose(fp);
+    if (line)
+        free(line);
+}
+
+ *
+ * User *load_users_from_file(char *filename)
+{
+    // TODO: import library from professor and read file name, importing users
+
+    return NULL;
+}
+
+Promotor *load_promoters_from_file(char *filename)
+{
+    // TODO: import library from professor and read file name, importing promoters
+
+    return NULL;
+}
+
+ */
 //list all items
 void exec_command_list() {
     printf("     > Executing the list all items command\n");
