+++ resources/file_users.txt
@@ -1,0 +1,2 @@
+Alex 123qwe. 1076
+Jorge joka123 600
