+++ shared/helpers/pipes.h
@@ -1,0 +1,3 @@
+
+#define BACKEND_FIFO_NAME "/tmp/BACKEND_COMMUNICATION_PIPE"
+#define FRONTEND_FIFO_NAME_BASE "/tmp/BACKEND_COMMUNICATION_PIPE"
